+++ lib/__init__.py
@@ -1,2 +1,2 @@
 # Wazuh Agent Manager Library
-__version__ = "1.8.0"
+__version__ = "1.9.0"